Add optional random flutter to the posting cooldown
Introduce a `cooldown_flutter_seconds` configuration key
(default 0, disabled). After each processed request the worker now
waits the full configured cooldown plus a fresh uniformly random
duration between zero and the flutter ceiling. Flutter only ever
lengthens the wait, so the guaranteed minimum spacing between posts
is preserved while the cadence stops being perfectly regular.

Use the `rand` crate for sampling and `saturating_add` so extreme
configurations cannot overflow. Move the worker unit tests to
`worker/tests.rs` to stay within the 400-line module cap.
